Fix flaky_attempt_count type and per-test retry override
Address PR review feedback:

- Change flaky_attempt_count from u32 to usize to match Zig's slice
  bounds requirements (used in flaky_attempts_buf[0..count] and as an
  array index).

- Change ParseArgumentsOptions.retry from u32 to ?u32 so that 'not
  provided' (null) is distinguishable from 'explicitly set to 0'. This
  allows per-test `{ retry: 0 }` to disable retries even when a global
  retry count is configured via --retry or bunfig. Previously, the
  fallback logic checked `retry == 0` which conflated the two cases.

## Test plan

zig:check-all passes on all platforms. Existing retry-flag and junit
reporter tests cover the runtime behavior.

+++ src/bun.js/test/Execution.zig
@@ -78,7 +78,7 @@ pub const ExecutionSequence = struct {
     test_entry: ?*ExecutionEntry,
     remaining_repeat_count: u32,
     remaining_retry_count: u32,
-    flaky_attempt_count: u32 = 0,
+    flaky_attempt_count: usize = 0,
     flaky_attempts_buf: [MAX_FLAKY_ATTEMPTS]FlakyAttempt = undefined,
     result: Result = .pending,
     executing: bool = false,
@@ -116,8 +116,10 @@ pub const ExecutionSequence = struct {
 
         pub fn failureType(this: FlakyAttempt) []const u8 {
             return switch (this.result) {
-                .fail_because_timeout, .fail_because_timeout_with_done_callback,
-                .fail_because_hook_timeout, .fail_because_hook_timeout_with_done_callback,
+                .fail_because_timeout,
+                .fail_because_timeout_with_done_callback,
+                .fail_because_hook_timeout,
+                .fail_because_hook_timeout_with_done_callback,
                 => "TimeoutError",
                 else => "AssertionError",
             };
